Changed how page names/titles work.
The controller for a page needs to set the title for the page. Most the time this title will just be a $translate() of some constant in the locale.json file, but occasionally it may be customized with information local to that controller such as search terms, user names, account info, etc.

Class that highlights active nav element is now applied by checking that the current state is the same as the state that nav links to. This way we don't need complex css to look for nth child under the pageant class. See header.css for details.