Optimized help component to generate the html.
It uses the pageName from the routes.config.json file as the key for translation. The translated pageName is displayed on the help modal as the nav text(on the left) and as the section headings.
routes.config.son also indicates how many Questions/Answers each help section will have. These values are stored in the locale.json files under help.pageName.